fix(codegen,vm): restore the llvm_codegen.cpp hunks dropped by the codegen extraction; VM manifold builtins stop shadowing core.manifold; defined shifts in the packed-string emitters (#592)
* fix(codegen,vm): restore the llvm_codegen.cpp hunks dropped by the codegen extraction; VM manifold builtins stop shadowing core.manifold; defined shifts in the packed-string emitters

The codegen extraction (bca3a6d2) was cut from 550771d4's lib/backend/llvm_codegen.cpp, so every change merged into that file afterwards was silently undone while the sibling modules and runtime headers kept the newer contracts: 893c7747 (shared Unicode classifier), 90203a8e (module privacy), 5e79cf31 (source-level symbol names in arity diagnostics), a324f35a (length-explicit interning), d9e0740c (fatal-error propagation), 53b8d264 (eshkol_arity_error_current call sites), eba46f59 (32-bit closure capture-count encoding), f34d846c (continuation state flags) and 6a721b16 (file-stat return type). The 31-capture parallel closure crash was the visible symptom: the lambda emitter wrote the 16-bit capture layout while the parallel dispatcher and CLOSURE_ENV_GET_NUM_CAPTURES read the 32-bit one.

Every dropped hunk is restored in place, audited commit by commit against the extracted modules. Also fixed from the same gate run: the VM builtin table aliased public core.manifold names to raw native ops with different representations and the bootstrap rule preferred the builtin over the module's definition (raw ops renamed to the -handle family, probe tests updated, VM prelude cache and language-surface manifest regenerated, documented builtin count 1,045 -> 1,043); the packed-string emitters in the VM compiler, hosted compiler and VM parser accumulate into uint64_t before the int64 constant (identical values, defined shifts); the header-layout comment copied into inc/eshkol/backend/llvm_codegen.h refers to eshkol_object_header_t instead of restating the layout.

* test(ml): exercise the canonical dataloader-has-next? predicate so the language-surface coverage gate holds at 1108/1108

(cherry picked from commit 4d44e70ba692a475ebf11cf7e83567b2ad6803f5)

* chore(coverage): ratchet the execution-backed language-coverage floor to 1109 with regenerated evidence (manifold handle builtins, canonical dataloader-has-next?)

* test(vm): exercise the process-spawn-argv-with-options alias in the VM system-surface regression

+++ lib/backend/vm_parser.c
@@ -1230,10 +1230,10 @@ static void compile_quote(FuncChunk* c, Node* datum) {
         int n_packs = (len + 7) / 8;
         chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(len)));
         for (int p = 0; p < n_packs; p++) {
-            int64_t pack = 0;
+            uint64_t pack = 0;
             for (int b = 0; b < 8 && p * 8 + b < len; b++)
-                pack |= ((int64_t)(unsigned char)datum->symbol[p * 8 + b]) << (b * 8);
-            chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL(pack)));
+                pack |= ((uint64_t)(unsigned char)datum->symbol[p * 8 + b]) << (b * 8);
+            chunk_emit(c, OP_CONST, chunk_add_const(c, INT_VAL((int64_t)pack)));
         }
         chunk_emit(c, OP_NATIVE_CALL,
                    ESHKOL_VM_PACKED_SYMBOL_FID_BASE + n_packs);
